+++ packages/talker_flutter/lib/src/ui/talker_monitor/talker_monitor_http_screen.dart
@@ -1,28 +1,47 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
+import 'package:share_plus/share_plus.dart';
 import 'package:talker_dio_logger/http_logs.dart';
+import 'package:talker_flutter/src/controller/controller.dart';
 import 'package:talker_flutter/src/extensions/iterable.dart';
 import 'package:talker_flutter/src/ui/theme/default_theme.dart';
+import 'package:talker_flutter/src/ui/widgets/actions_bottom_sheet.dart';
 import 'package:talker_flutter/talker_flutter.dart';
 
-class TalkerMonitorHttpScreen extends StatelessWidget {
+class TalkerMonitorHttpScreen extends StatefulWidget {
   const TalkerMonitorHttpScreen({
     Key? key,
     required this.talker,
-    required this.theme,
+    required this.talkerScreenTheme,
   }) : super(key: key);
 
   final TalkerInterface talker;
-  final TalkerScreenTheme theme;
+  final TalkerScreenTheme talkerScreenTheme;
+
+  @override
+  State<TalkerMonitorHttpScreen> createState() =>
+      _TalkerMonitorHttpScreenState();
+}
+
+class _TalkerMonitorHttpScreenState extends State<TalkerMonitorHttpScreen> {
+  var _expandRequestLogs = true;
+  final _controller = TalkerScreenController();
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: const Text('Talker Http Monitor'),
+        actions: [
+          IconButton(
+            onPressed: () => _showBottomSheet(context),
+            icon: const Icon(Icons.more_vert_rounded),
+          ),
+        ],
       ),
-      backgroundColor: theme.backgroudColor,
+      backgroundColor: widget.talkerScreenTheme.backgroudColor,
       body: TalkerHistoryBuilder(
-        talker: talker,
+        talker: widget.talker,
         builder: (context, data) {
           final reqRespPairMap = _mapHttpLogsToPairs(data);
           return CustomScrollView(
@@ -37,6 +56,9 @@ class TalkerMonitorHttpScreen extends StatelessWidget {
                         TalkerDataCard(
                           data: pair.key,
                           margin: EdgeInsets.zero,
+                          expanded: _expandRequestLogs,
+                          onTap: () =>
+                              _copyTalkerDataItemText(context, pair.key),
                         ),
                         if (pair.value != null) ...[
                           Container(
@@ -55,7 +77,12 @@ class TalkerMonitorHttpScreen extends StatelessWidget {
                               ),
                             ),
                           ),
-                          TalkerDataCard(data: pair.value!),
+                          TalkerDataCard(
+                            data: pair.value!,
+                            expanded: _expandRequestLogs,
+                            onTap: () =>
+                                _copyTalkerDataItemText(context, pair.value!),
+                          ),
                         ],
                         const SizedBox(height: 10),
                       ],
@@ -68,6 +95,20 @@ class TalkerMonitorHttpScreen extends StatelessWidget {
           );
         },
       ),
+    );
+  }
+
+  void _copyTalkerDataItemText(BuildContext context, TalkerDataInterface data) {
+    final text = data is FlutterTalkerDataInterface
+        ? data.generateFlutterTextMessage()
+        : data.generateTextMessage();
+    Clipboard.setData(ClipboardData(text: text));
+    _showSnackBar(context, 'Log item is copied in clipboard');
+  }
+
+  void _showSnackBar(BuildContext context, String text) {
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(content: Text(text)),
     );
   }
 
@@ -91,5 +132,64 @@ class TalkerMonitorHttpScreen extends StatelessWidget {
       return MapEntry(e, null);
     }).toList();
     return reqRespPairMap;
+  }
+
+  void _showBottomSheet(BuildContext context) async {
+    await showModalBottomSheet(
+      context: context,
+      isScrollControlled: true,
+      backgroundColor: Colors.transparent,
+      builder: (context) {
+        return ActionsBottomSheet(
+          actions: [
+            BottomSheetAction(
+              onTap: _controller.toggleLogOrder,
+              title: 'Reverse logs',
+              icon: Icons.swap_vert,
+            ),
+            BottomSheetAction(
+              onTap: () => _copyAllLogs(context),
+              title: 'Copy all logs',
+              icon: Icons.copy,
+            ),
+            BottomSheetAction(
+              onTap: () => _toggleRequestLogsExpanded(!_expandRequestLogs),
+              title: _expandRequestLogs ? 'Collapse logs' : 'Expand logs',
+              icon: _expandRequestLogs
+                  ? Icons.visibility_outlined
+                  : Icons.visibility_off_outlined,
+            ),
+            BottomSheetAction(
+              onTap: _shareLogsInFile,
+              title: 'Share logs file',
+              icon: Icons.ios_share_outlined,
+            ),
+          ],
+          talkerScreenTheme: widget.talkerScreenTheme,
+        );
+      },
+    );
+  }
+
+  void _toggleRequestLogsExpanded(bool value) {
+    setState(() => _expandRequestLogs = value);
+  }
+
+  Future<void> _shareLogsInFile() async {
+    final path = await _controller.saveLogsInFile(
+      _httpLogs.text,
+    );
+    // ignore: deprecated_member_use
+    await Share.shareFilesWithResult([path]);
+  }
+
+  List<TalkerDataInterface> get _httpLogs => widget.talker.history
+      .where((e) =>
+          e is HttpRequestLog || e is HttpErrorLog || e is HttpResponseLog)
+      .toList();
+
+  void _copyAllLogs(BuildContext context) {
+    Clipboard.setData(ClipboardData(text: _httpLogs.text));
+    _showSnackBar(context, 'All logs copied in buffer');
   }
 }