+++ packages/talker_flutter/lib/src/ui/talker_monitor/widgets/widgets.dart
@@ -1,0 +1,1 @@
+export 'talker_monitor_item.dart';