+++ utils/pub_get_all.sh
@@ -15,3 +15,15 @@ flutter pub get
 echo "talker_dio_logger"
 cd ../talker_dio_logger
 flutter pub get
+
+echo "talker_bloc_logger"
+cd ../talker_bloc_logger
+flutter pub get
+
+echo "shop_app_example"
+cd ../../examples/shop_app_example
+flutter pub get
+
+echo "custom_logs_example"
+cd ../../examples/custom_logs_example
+flutter pub get