+++ packages/talker_flutter/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.3.0
+- **FEAT**: Add filters and fast actions for http calls monitor
+- **FEAT**: Expanded size of log item click zone
+
 # 2.2.1
 - **FIX**: Fix color reset in console